feat(product): add QuantityButtons and AddToCartButton components for improved product interaction feat(product): create ProductDetailsSkeleton component for loading state representation fix(index.ts): export new product components to make them available for use in other parts of the application
