+++ client/src/components/product/ProductDetailsSkeleton.tsx
@@ -1,0 +1,1 @@
+export function ProductDetailsSkeleton() {}
