+++ client/src/routes/menu.product.$productId.tsx
@@ -1,7 +1,9 @@
 // Imports
-import { createFileRoute, Link } from '@tanstack/react-router';
+import { createFileRoute, Link, useParams } from '@tanstack/react-router';
 import productImage from '../assets/home-hero.jpeg';
+import { QuantityButtons, AddToCartButton } from '../components';
 import { useState } from 'react';
+// import { useGetProductById } from '../hooks/useProductsHooks';
 
 export const Route = createFileRoute('/menu/product/$productId')({
   component: ProductDetails,
@@ -10,15 +12,13 @@ export const Route = createFileRoute('/menu/product/$productId')({
 function ProductDetails() {
   const [quantity, setQuantity] = useState(1);
 
-  function handleIncrement() {
-    setQuantity(quantity + 1);
-  }
+  const productId = useParams({
+    from: '/menu/product/$productId',
+    select: (params) => params.productId,
+  });
 
-  function handleDecrement() {
-    if (quantity > 1) {
-      setQuantity(quantity - 1);
-    }
-  }
+  // const { data: product, isPending } = useGetProductById(productId);
+  console.log(productId);
 
   return (
     <div className="mb-8 lg:mx-32 xl:mx-72">
@@ -47,32 +47,8 @@ function ProductDetails() {
             Tender chicken or pork marinated in soy sauce, vinegar, and garlic,
             slow-cooked to perfection.
           </p>
-          <div
-            id="cart-counter"
-            className="mt-4 flex items-center justify-center"
-          >
-            <button
-              id="decrement"
-              className="btn btn-outline"
-              onClick={handleDecrement}
-            >
-              -
-            </button>
-            <input
-              type="number"
-              value={quantity}
-              className="w-16 appearance-none bg-inherit text-center [&::-webkit-inner-spin-button]:appearance-none [&::-webkit-outer-spin-button]:appearance-none"
-            />
-            <button
-              id="increment"
-              className="btn btn-outline"
-              onClick={handleIncrement}
-            >
-              +
-            </button>
-          </div>
-
-          <button className="btn btn-outline mt-6 w-full">Add to Cart</button>
+          <QuantityButtons quantity={quantity} setQuantity={setQuantity} />
+          <AddToCartButton quantity={quantity} />
         </div>
       </div>
     </div>
